Introducing application cache
Loading app's icon is an expensive task.
That's why when you do a fast scrolling through apps list, the UI is jerky.
If the cache is not populated yet, it will create a cache of app's icon and app's name only once, unless the apps list needs to be manually refreshed (by pulling-down the apps list of package disabler or mobile restricter).
Creating cache takes time, you will see a progress dialog.