DomainsManagement: Fix list-jumping behavior issue
There was a fundamental issue with the way how to update the content of the list (ListView <-> ListAdapter).
For every change on the list item, the list gets notified and the content of the list gets recreated instead just gets updated.
This causes the list to jump to the first item for every change.