Fix issue when domain rules are disabled although they are enabled
This can happen when the total domains count more than Knox firewall can handle.
We apply the domains in chunks but we get the total domains count with a single function.
Since in this case, we cannot get the total domains count from Knox firewall, we will get it from database.
This number could higher than the actual number in Knox firewall because to get the count for total unique domains is slow.
For the domain limit below 15000, the count is retrieved from Knox firewall.