Handle app install/uninstall event differently
On Oreo, the explicit broadcast receiver is not allowed anymore.
This causes the app list is not updated when apps are installed/uninstalled.
On Nougat, this works fine.

To mitigate this restriction, the app list will be refreshed automatically when opening an app while caching apps.
This is fast because it is done in parallel by utilizing half of the available cpus.
Refreshing the app list with pull-down gesture uses the same code as caching apps which is now faster than before.

Refreshing the app list means to delete the database and applied your setup for disabled apps, etc.
If refreshing the app list went wrong, you need to restore the database. So, make sure to create a database backup once you setup everything.

By using this strategy, we don't need the app receiver anymore - This receiver is listening on app install/uninstall.
So, now Adhell3 is more lightweight with one receiver less.