App Disabler toggle
Before this change, when you want to enable all "disabled apps", you need to do it with "Enabling all apps".
With this, you will lose your setup of disabled apps as all apps are toggled to be enabled.
The only way to bring your setup back (which apps were disabled) is through restore database.

Now you can enable all "disabled apps" while still retaining your disabled apps setup.
If the toggle is enabled, changes to the app (enabling or disabling it) is applied immediately as before.