Implement less restrictive blocking Knox does not seem to consider prefixes that do not contain letters to be valid.
9.9gag.com -> Fail
9.9gag.com* -> Fail
*9.9gag.com -> Pass
t.9gag.com -> Pass
t9.9gag.com -> Pass

34342342343.com -> Fail
34342342343.com* -> Fail
*34342342343.com -> Pass
t34342342343.com -> Pass

So, we now say: If url is a wildcard, return as is. Otherwise, if it has letters, return as is || if it doesn't have letters, add a * prefix.

Will need to raise a separate issue as the url reporting is wrong. Presumably it reads straight from the DB.