Check existing domains in the database for wildcard prefix
After commit c788db6fd, we are storing the valid Knox domain in the database.
The process to prefix the domain with the wildcard is only done when a new domain for blacklist/whitelist or host provider is added.
Therefore, for the domains that are already stored in the database, we need to prefixing them.
This will only run once.