Migration to Knox SDK
* Knox SDK license key is now needed
It contains two keys: Samsung Knox License (SKL) key and Backwards-compatible key
For devices with Knox 2.8 and newer, only SKL key is needed.
For devices with Knox 2.7.1 or earlier, the backwards-compatible key additionally to the SKL key is needed.

* Support only Knox version 2.6 - 3.1 (API level 19 - 25)
Knox version 2.6 is equivalent to Standard SDK version 5.6. See this site for the version mapping: https://seap.samsung.com/sdk/knox-version-mapping
The API to support Knox version 2.5 and earlier is not available anymore in Knox SDK. Hence, I drop these versions.

* App permission behavior changes
With Knox SDK, it is only possible to disable app permission that has protection-level of 'dangerous'.
If a permission is disabled, other permissions belong to the same group will also be disabled. E.g. it is not possible anymore to just disable WRITE_EXTERNAL_STORAGE without disabling READ_EXTERNAL_STORAGE.
This functionality is basically the same if you disable app permission in Android Settings - Apps - [AppName] - Permissions. You only see the permission groups.

* Not using local database anymore
Starting with 3.1.1.x, the Adhell3 database won't be created in internal storage anymore.
This was a convenience feature when most of the people need to re-install Adhell3 and each time with different package name.
Since you can compile it by yourself, you can just assign the same package name so that your existing Adhell3 will be updated.