Don't update the home screen twice
Switching between menu at the bottom tab causes the selected tab to be recreated.
When the tab is recreated, the onResume is also executed which causing double call of updateUserInterface().
As consequence, when Adhell is in the background because another app is active in foreground, home screen won't be updated. But this is less concern than having double call of updateUserInterface().