Protect Adhell3 with a password
The password is hashed with PBKDF2 to give more protection against brute force attacks if the attacker gets the password hash.
The password hash is stored in the application's data partition which can only be accessible with a root privilege.
If you forget the password, you need to uninstall Adhell3.