Ability to change license key
Before changing the license key, the old one needs to be deactivated first.
Deactivating key means to remove the license from the device. It doesn't mean to revoke the license itself.
The validity of the license key is controlled in seap.samsung.com.

Note: If you deactivating the license, the activation dialog cannot be dismissed until a new key can be activated.